requirements.txt
pyserial==3.5
rich==13.4.0
pyfiglet==0.8.post1

